Completely changed how constants were handled. I didn't like how there would be a whole object being made for each decaying stat all reading from the same file.
To optimize it as best I could, I turned the Constants class into a
config reader of sorts and sent a HashMap to the main GUI class and
passed the variables around as needed.

DOWNSIDE/TODO: Right now, all of the strings for each constant read from
the file are hardcoded wherever they're used. I want to make it more
managable for if a variable name changes. Make a string for each
constant and change it when necessary.